requirements.txt
# Core Deep Learning
tensorflow>=2.10.0,<3.0.0
keras>=2.10.0

# Computer Vision
opencv-python>=4.6.0
mediapipe>=0.10.0
Pillow>=9.0.0

# Scientific Computing
numpy>=1.21.0
scikit-learn>=1.0.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.12.0

# GUI (tkinter is usually included with Python, but listed for completeness)
# tkinter - included in Python standard library

# Progress bars
tqdm>=4.64.0

# Image processing utilities
# PIL/Pillow already included above

# Optional but recommended
# pandas>=1.3.0  # For data manipulation if needed
# jupyter>=1.0.0  # For notebook development

